Add argument control to running SDM scripts
The decision to run a few (8) species or all (~150) is toggled via a
logical (all_insects) inside respective run-all-SDM-<method>-scripts.R
files. To better accommodate toggling on/off, add functionality that
will look for an -a flag in the command arguments. If flag is set, it
runs all insects (and hosts). If flag is not passed in call to script,
it uses whatever value is assigned to all_insects in the R script.
Note this functionality is only used when calling the R script from a
bash command line, e.g.:
$ Rscript run-all-SDM-<method>-scripts.R -a # runs all insects
$ Rscript run-all-SDM-<method>-scripts.R    # run dictated by all_insects value